Inserido a data de criação e permissão de nível de acesso

+++ php_action/excluirN.php
@@ -3,9 +3,13 @@
 
     //Exclui o nível de acesso selecionado
     $id = $_POST['id'];
-    if($id > 0){
+    $idUserLogado = $_POST['idUserLogado'];
+
+    if($id > 0 && $idUserLogado == 1){
         $sql = "DELETE FROM nivel_acesso WHERE (Id_Nivel_Acesso = '{$id}')";
         $conn->query($sql);
+    }else{
+        echo("Somente administradores podem excluir um usuário");
     }
     $conn->close();
     header("Location:../index.php");
