+++ php_action/excluirU.php
@@ -3,9 +3,13 @@
 
     //Exclui o usuario selecionado
     $id = $_POST['id'];
-    if($id > 0){
+    $idUserLogado = $_POST['idUserLogado'];
+
+    if($id > 0 && $idUserLogado == 1){
         $sql = "DELETE FROM usuario WHERE (Id_Usuario = '{$id}')";
         $conn->query($sql);
+    }else{
+        echo("Somente administradores podem excluir um nível");
     }
     $conn->close();
     header("Location:../index.php");
